Added option to allow manual response
The origins of this change came from trying to use auth0.js SDK to allow the use of a popup while passing the callback to nextjs-auth0 SDK using api/routes to handle the session creation.  To close the popup ‘popup.callback()’ must be called after the user login has completed.  This can be done by using redirectTo and passing url to a simple page that contains this code.  However since this code is quite simple, I wanted to simply send the page in the response directly from the callback api endpoint as such:

res.send(`<html>
      <body>
        <script src="https://cdn.auth0.com/js/auth0/9.11/auth0.min.js"></script>
        <script type="text/javascript">
          var webAuth = new auth0.WebAuth({
            domain:       '${process.env.AUTH0_DOMAIN}',
            clientID:     '${process.env.AUTH0_CLIENT_ID}',
          });
          webAuth.popup.callback();
        </script>
      </body>
    </html>`);